Revert "Split out shared stuff from navs to navbase"
This reverts commit b939c719270d3eead7b96fde012fcddd82230605.
